requirements_gpt.txt
fastapi
uvicorn[standard]
pillow
opencv-python
pytesseract
pykakasi
fugashi
unidic-lite
pyttsx3
python-multipart
